# docs/openapi.yaml
openapi: 3.1.0
info:
  title: InkFlow CRM API
  description: |
    REST API for InkFlow tattoo studio platform.

    Base path: /api (server context-path).

    Response envelope: ApiResponse<T> - see docs/API_FORMAT.md.

    Auth:
    - CRM (admin mobile / web): Supabase JWT in Authorization Bearer header.
      Custom claims: tenant_id, role, location_ids (via Supabase hook).
    - Consumer (B2C app): Supabase JWT for /public/consumer/** routes.
    - Public reads: GET /public/catalog/** and GET /public/artists/** - no token.
  contact:
    name: InkFlow
    email: dev@inkflow.studio
  license:
    name: Proprietary
  version: 1.0.0
servers:
- url: /api
  description: Current host (context-path included)
paths:
...
  /staff/{id}/services:
    get:
      tags:
      - CRM Â· Staff
      operationId: getStaffServices
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListStaffServiceDto"
    put:
      tags:
      - CRM Â· Staff
      operationId: updateStaffServices
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffServicesRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListStaffServiceDto"
  /staff/{id}/services/{serviceId}:
    put:
      tags:
      - CRM Â· Staff
      operationId: updateStaffServicePricing
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: serviceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddStaffServiceRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseStaffServiceDto"
    post:
      tags:
      - CRM Â· Staff
      operationId: addServiceToStaff
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: serviceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddStaffServiceRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseStaffServiceDto"
    delete:
      tags:
      - CRM Â· Staff
      operationId: removeServiceFromStaff
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: serviceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /staff/{id}/schedule:
    put:
      tags:
      - CRM Â· Staff
      operationId: updateSchedule
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScheduleRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /staff/{id}/faq:
    get:
      tags:
      - CRM Â· Staff
      operationId: getFaq
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListStaffFaqDto"
    put:
      tags:
      - CRM Â· Staff
      operationId: upsertFaq
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertFaqRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListStaffFaqDto"
...
  /staff:
    get:
      tags:
      - CRM Â· Staff
      operationId: getAllStaff
      parameters:
      - name: pageRequest
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/PageRequest"
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: role
        in: query
        required: false
        schema:
          type: string
      - name: locationId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: accountStatus
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListStaffDto"
    post:
      tags:
      - CRM Â· Staff
      operationId: createStaff
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseStaffDto"
...
  /staff/{id}/reactivate:
    post:
      tags:
      - CRM Â· Staff
      operationId: reactivateStaff
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /staff/{id}/deactivate:
    post:
      tags:
      - CRM Â· Staff
      operationId: deactivateStaff
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeactivateStaffRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /staff/invite:
    post:
      tags:
      - CRM Â· Staff
      operationId: inviteStaff
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteStaffRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringString"
  /staff/accept-invite:
    post:
      tags:
      - CRM Â· Staff
      operationId: acceptInvite
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptInviteRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseStaffDto"
...
  /staff/{id}:
    get:
      tags:
      - CRM Â· Staff
      operationId: getStaff
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseStaffDetailDto"
    delete:
      tags:
      - CRM Â· Staff
      operationId: deleteStaff
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
    patch:
      tags:
      - CRM Â· Staff
      operationId: updateStaff
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseStaffDto"
...
  /staff/{id}/future-appointments-count:
    get:
      tags:
      - CRM Â· Staff
      operationId: getFutureAppointmentsCount
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringInteger"
  /staff/invite/info/{token}:
    get:
      tags:
      - CRM Â· Staff
      operationId: getInviteInfo
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseInviteInfoDto"
...
  /leaves/staff/{staffId}:
    get:
      tags:
      - CRM Â· Leave
      operationId: getLeavesByStaffId
      parameters:
      - name: staffId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: startDate
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: endDate
        in: query
        required: false
        schema:
          type: string
          format: date
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListLeaveRequestDto"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ErrorDto"
...
    PaginationDto:
      type: object
      properties:
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
...
    ServiceAssignment:
      type: object
      properties:
        serviceId:
          type: string
          format: uuid
        customPrice:
          type: number
        customDuration:
          type: integer
          format: int32
      required:
      - serviceId
    UpdateStaffServicesRequest:
      type: object
      properties:
        services:
          type: array
          items:
            $ref: "#/components/schemas/ServiceAssignment"
      required:
      - services
    ApiResponseListStaffServiceDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/StaffServiceDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    StaffServiceDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        pricingType:
          type: string
        basePrice:
          type: number
        customPrice:
          type: number
        baseDuration:
          type: integer
          format: int32
        customDuration:
          type: integer
          format: int32
        color:
          type: string
        isActive:
          type: boolean
    AddStaffServiceRequest:
      type: object
      properties:
        customPrice:
          type: number
        customDuration:
          type: integer
          format: int32
    ApiResponseStaffServiceDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/StaffServiceDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    ScheduleEntry:
      type: object
      properties:
        dayOfWeek:
          type: string
          pattern: ^(monday|tuesday|wednesday|thursday|friday|saturday|sunday)$
        isWorking:
          type: boolean
        startTime:
          type: string
          pattern: "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
        endTime:
          type: string
          pattern: "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
      required:
      - dayOfWeek
      - isWorking
    UpdateScheduleRequest:
      type: object
      properties:
        schedule:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleEntry"
      required:
      - schedule
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    FaqItem:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
      required:
      - answer
      - question
    UpsertFaqRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FaqItem"
    ApiResponseListStaffFaqDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/StaffFaqDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    StaffFaqDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        question:
          type: string
        answer:
          type: string
        sortOrder:
          type: integer
          format: int32
...
    CreateStaffRequest:
      type: object
      properties:
        firstName:
          type: string
          maxLength: 50
          minLength: 1
        lastName:
          type: string
          maxLength: 50
          minLength: 1
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          pattern: ^(owner|admin|artist)$
        calendarColor:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
        specialization:
          type: array
          items:
            type: string
        bio:
          type: string
        locationIds:
          type: array
          items:
            type: string
            format: uuid
      required:
      - calendarColor
      - email
      - firstName
      - lastName
      - locationIds
      - role
    ApiResponseStaffDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/StaffDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    StaffDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        avatar:
          type: string
        role:
          type: string
        calendarColor:
          type: string
        specialization:
          type: array
          items:
            type: string
        portfolioImages:
          type: array
          items:
            type: string
        status:
          type: string
        accountStatus:
          type: string
        locationIds:
          type: array
          items:
            type: string
            format: uuid
        salaryType:
          type: string
        salaryRate:
          type: number
        bankDetails:
          type: string
        position:
          type: string
        birthday:
          type: string
          format: date
        taxId:
          type: string
        iban:
          type: string
        bankCard:
          type: string
        availableForOnlineBooking:
          type: boolean
        isServiceProvider:
          type: boolean
        instagram:
          type: string
        hourlyRate:
          type: number
        dontDoList:
          type: array
          items:
            type: string
        bio:
          type: string
        createdAt:
          type: string
          format: date-time
...
    DeactivateStaffRequest:
      type: object
      properties:
        cancelFutureAppointments:
          type: boolean
    InviteStaffRequest:
      type: object
      properties:
        email:
          type: string
        role:
          type: string
          pattern: ^(admin|artist)$
        calendarColor:
          type: string
        isServiceProvider:
          type: boolean
        locationIds:
          type: array
          items:
            type: string
            format: uuid
      required:
      - email
      - locationIds
      - role
    ApiResponseMapStringString:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          additionalProperties:
            type: string
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    AcceptInviteRequest:
      type: object
      properties:
        token:
          type: string
        firstName:
          type: string
          maxLength: 50
          minLength: 1
        lastName:
          type: string
          maxLength: 50
          minLength: 1
        phone:
          type: string
        authUserId:
          type: string
      required:
      - authUserId
      - firstName
      - lastName
      - token
...
    LeaveRequestDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        staffId:
          type: string
          format: uuid
        staffName:
          type: string
        staffAvatar:
          type: string
        leaveType:
          type: string
        status:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        daysCount:
          type: integer
          format: int32
        reason:
          type: string
        notes:
          type: string
        approvedById:
          type: string
          format: uuid
        approvedByName:
          type: string
        approvedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    UpdateStaffRequest:
      type: object
      properties:
        firstName:
          type: string
          maxLength: 50
          minLength: 1
        lastName:
          type: string
          maxLength: 50
          minLength: 1
        email:
          type: string
        phone:
          type: string
        avatar:
          type: string
        calendarColor:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
        specialization:
          type: array
          items:
            type: string
        portfolioImages:
          type: array
          items:
            type: string
        bio:
          type: string
        locationIds:
          type: array
          items:
            type: string
            format: uuid
        status:
          type: string
          pattern: ^(working|on_vacation|sick_leave|fired)$
        salaryType:
          type: string
          pattern: ^(none|fixed|percent)$
        salaryRate:
          type: number
          minimum: 0.0
        bankDetails:
          type: string
        position:
          type: string
        birthday:
          type: string
          format: date
        taxId:
          type: string
        iban:
          type: string
        bankCard:
          type: string
        availableForOnlineBooking:
          type: boolean
        isServiceProvider:
          type: boolean
        instagram:
          type: string
        hourlyRate:
          type: number
        dontDoList:
          type: array
          items:
            type: string
...
    PageRequest:
      type: object
      properties:
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        sortBy:
          type: string
        sortDir:
          type: string
...
    ApiResponseListStaffDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/StaffDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    ApiResponseStaffDetailDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/StaffDetailDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
...
    StaffDetailDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        avatar:
          type: string
        role:
          type: string
        calendarColor:
          type: string
        specialization:
          type: array
          items:
            type: string
        portfolioImages:
          type: array
          items:
            type: string
        bio:
          type: string
        status:
          type: string
        accountStatus:
          type: string
        locations:
          type: array
          items:
            $ref: "#/components/schemas/LocationDto"
        schedule:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleDto"
        stats:
          $ref: "#/components/schemas/StaffStatsDto"
        salaryType:
          type: string
        salaryRate:
          type: number
        bankDetails:
          type: string
        isServiceProvider:
          type: boolean
        instagram:
          type: string
        hourlyRate:
          type: number
        dontDoList:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ApiResponseMapStringInteger:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          additionalProperties:
            type: integer
            format: int32
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    ApiResponseInviteInfoDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/InviteInfoDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    InviteInfoDto:
      type: object
      properties:
        email:
          type: string
        role:
          type: string
        expiresAt:
          type: string
          format: date-time
        expired:
          type: boolean
        accepted:
          type: boolean
...
    ApiResponseListLeaveRequestDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/LeaveRequestDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
...
  securitySchemes:
    CrmBearer:
      type: http
      description: Supabase JWT for CRM staff (OWNER/ADMIN/ARTIST).
      scheme: bearer
      bearerFormat: JWT
    ConsumerBearer:
      type: http
      description: Supabase JWT for B2C consumer app users.
      scheme: bearer
      bearerFormat: JWT